# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GUM Internships API
  version: 1.0.0
  description: >
    MVP API contract. Every response uses the envelope {success, data, error, meta}.
    DB is snake_case; this API is camelCase. Auth = Bearer JWT (access 15m, refresh 30d).
    Pagination on list endpoints: ?page=&limit= (default 20, max 100).
    Endpoints marked "Public" need no token. x-required-roles lists allowed roles.
servers:
  - url: https://api.gum-internships.example.com/v1
    description: Production
  - url: https://staging-api.gum-internships.example.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /tickets:
    post:
      tags: [Support]
      summary: Raise a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, subject, description]
              properties:
                category: { type: string, enum: [payment, content, technical, certificate, other] }
                subject: { type: string }
                description: { type: string }
                internshipId: { type: integer, format: int64 }
                attachments: { type: array, items: { type: string } }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Ticket' } }
  /tickets/me:
    get:
      tags: [Support]
      summary: My tickets
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Ticket' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
  /tickets/{ticketId}:
    get:
      tags: [Support]
      summary: Ticket detail (owner only)
      parameters:
        - { name: ticketId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Ticket' } }

  # ============================================================ ADMIN
...
  /admin/instructors:
    get:
      tags: [Admin]
      summary: Instructor profiles (filter kycStatus)
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: kycStatus, in: query, schema: { type: string, enum: [pending, submitted, approved, rejected] } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/InstructorProfile' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
  /admin/instructors/{profileId}/kyc-decision:
    post:
      tags: [Admin]
      summary: Approve/reject KYC (approval grants instructor role, sets share %)
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: profileId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                reason: { type: string, description: Required when rejected }
                revenueSharePercent: { type: number, minimum: 0, maximum: 100 }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
...
  /admin/internships/{internshipId}/decision:
    post:
      tags: [Admin]
      summary: Publish/reject a pending internship
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: internshipId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [published, rejected] }
                reason: { type: string, description: Required when rejected }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
...
  /admin/users/{userId}/status:
    patch:
      tags: [Admin]
      summary: Suspend/reactivate user
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, suspended] }
                reason: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
  /admin/users/{userId}/roles:
    post:
      tags: [Admin]
      summary: Grant role (super_admin only)
      x-required-roles: [super_admin]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [student, instructor, moderator, finance_admin, support, super_admin] }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
  /admin/users/{userId}/roles/{role}:
    delete:
      tags: [Admin]
      summary: Revoke role (super_admin only)
      x-required-roles: [super_admin]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: role, in: path, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
...
  /admin/audit-logs:
    get:
      tags: [Admin]
      summary: Search audit trail
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: actorId, in: query, schema: { type: integer, format: int64 } }
        - { name: action, in: query, schema: { type: string } }
        - { name: entityType, in: query, schema: { type: string } }
        - { name: entityId, in: query, schema: { type: integer, format: int64 } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/AuditLog' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
  /admin/tickets:
    get:
      tags: [Admin]
      summary: Support queue (filter status/priority/category)
      x-required-roles: [support, moderator, finance_admin, super_admin]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, in_progress, resolved, closed] } }
        - { name: priority, in: query, schema: { type: string, enum: [low, medium, high, urgent] } }
        - { name: category, in: query, schema: { type: string, enum: [payment, content, technical, certificate, other] } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Ticket' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
  /admin/tickets/{ticketId}:
    patch:
      tags: [Admin]
      summary: Assign / set priority / resolve with note
      x-required-roles: [support, moderator, finance_admin, super_admin]
      parameters:
        - { name: ticketId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [open, in_progress, resolved, closed] }
                priority: { type: string, enum: [low, medium, high, urgent] }
                assignedTo: { type: integer, format: int64 }
                resolutionNote: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
...
  /admin/cms/banners:
    get:
      tags: [Admin]
      summary: All banners
      x-required-roles: [moderator, super_admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Banner' } }
    post:
      tags: [Admin]
      summary: Create banner
      x-required-roles: [moderator, super_admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, imageUrl, placement]
              properties:
                title: { type: string }
                imageUrl: { type: string }
                linkUrl: { type: string }
                placement: { type: string, enum: [home_hero, home_strip, category_page] }
                displayOrder: { type: integer }
                startsAt: { type: string, format: date-time }
                endsAt: { type: string, format: date-time }
      responses:
        '201': { $ref: '#/components/responses/OkMessage' }
  /admin/cms/banners/{bannerId}:
    patch:
      tags: [Admin]
      summary: Update/deactivate banner
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: bannerId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
    delete:
      tags: [Admin]
      summary: Delete banner
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: bannerId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
  /admin/cms/pages:
    get:
      tags: [Admin]
      summary: All CMS pages
      x-required-roles: [moderator, super_admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/CmsPage' } }
    post:
      tags: [Admin]
      summary: Create page
      x-required-roles: [moderator, super_admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, title, contentMd]
              properties:
                slug: { type: string }
                title: { type: string }
                contentMd: { type: string }
                metaTitle: { type: string }
                metaDescription: { type: string }
                isPublished: { type: boolean, default: false }
      responses:
        '201': { $ref: '#/components/responses/OkMessage' }
  /admin/cms/pages/{pageId}:
    patch:
      tags: [Admin]
      summary: Update/publish page
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: pageId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    OkMessage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - properties:
                  data:
                    type: object
                    properties: { message: { type: string } }
...
  schemas:
    Envelope:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, const: true }
        data: {}
        error: { type: 'null' }
        meta: { type: object }
...
    PaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    InstructorProfile:
      type: object
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        instructorType: { type: string, enum: [internal, external] }
        bio: { type: [string, 'null'] }
        expertise: { type: array, items: { type: string } }
        kycStatus: { type: string, enum: [pending, submitted, approved, rejected] }
        agreementStatus: { type: string, enum: [pending, sent, signed] }
        revenueSharePercent: { type: number }
        bankAccountLast4: { type: [string, 'null'] }
        rejectionReason: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
...
    Ticket:
      type: object
      properties:
        id: { type: integer, format: int64 }
        ticketNo: { type: string }
        category: { type: string, enum: [payment, content, technical, certificate, other] }
        subject: { type: string }
        description: { type: string }
        status: { type: string, enum: [open, in_progress, resolved, closed] }
        priority: { type: string, enum: [low, medium, high, urgent] }
        resolutionNote: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    Banner:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        imageUrl: { type: string }
        linkUrl: { type: [string, 'null'] }
        placement: { type: string, enum: [home_hero, home_strip, category_page] }
        displayOrder: { type: integer }
    CmsPage:
      type: object
      properties:
        id: { type: integer, format: int64 }
        slug: { type: string }
        title: { type: string }
        contentMd: { type: string }
        metaTitle: { type: [string, 'null'] }
        metaDescription: { type: [string, 'null'] }
    AuditLog:
      type: object
      properties:
        id: { type: integer, format: int64 }
        actorId: { type: [integer, 'null'], format: int64 }
        actorRole: { type: [string, 'null'] }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: [integer, 'null'], format: int64 }
        beforeData: { type: [object, 'null'] }
        afterData: { type: [object, 'null'] }
        createdAt: { type: string, format: date-time }